adds switch to omit source code parsing for documentation

+++ volumes/shared/docs/build.sh
@@ -1,13 +1,26 @@
 #!/bin/bash
+# For copyright and license terms, see COPYRIGHT.rst (top level of repository)
+# Repository: https://github.com/C3S/collecting_society_docker
+#
+# Usage: ./docs-build [--no-autoapi]
+#
+# Options:
+#     --no-autoapi: don't parse the modules
 
-rm -rf build/doctrees
-rm -rf build/html
-rm -rf source/generated
+NOAUTOAPI=false
+for i in "$@"; do
+    if [ $i = "--no-autoapi" ]; then NOAUTOAPI=true; continue; fi
+done
 
-sphinx-apidoc -f -o source/generated --no-toc ../src/collecting_society
-sphinx-apidoc -f -o source/generated --no-toc ../src/collecting_society_web
-sphinx-apidoc -f -o source/generated --no-toc ../src/collecting_society_worker
-sphinx-apidoc -f -o source/generated --no-toc ../src/portal_web
+if ! $NOAUTOAPI; then
+    rm -rf build/doctrees
+    rm -rf build/html
+    rm -rf source/generated
+    sphinx-apidoc -f -o source/generated --no-toc ../src/collecting_society
+    sphinx-apidoc -f -o source/generated --no-toc ../src/collecting_society_web
+    sphinx-apidoc -f -o source/generated --no-toc ../src/collecting_society_worker
+    sphinx-apidoc -f -o source/generated --no-toc ../src/portal_web
+fi
 
 # echo ".. title:: Collecting Society Docker Setup" > source/generated/collecting_society_docker_README.rst
 cp /collecting_society_docker/README.rst source/generated/collecting_society_docker_README.rst
